% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{getAgeMax}
\alias{getAgeMax}
\title{\code{getAgemax} -- Get the maximum age omega of a discrete life table}
\usage{
getAgeMax(object)
}
\arguments{
\item{object}{A life table object}
}
\value{
a numeric coreponding to maximum ages that insuree could be.
}
\description{
This method gets you the maximum age from a given life table. The life table have to be discrete.
}
\examples{
getAgeMax(th002d)
getAgeMax(tableGPDId)
}
\seealso{
\code{\link{Lx}}, \code{\link{getAgeMax}}, \code{\linkS4class{TableMortDiscrete}}, \code{\link{getTable}}
}
