% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{getCdf}
\alias{getCdf}
\title{\code{getCdf} -- Get the Cdf from a continous life table}
\usage{
getCdf(object, ...)
}
\arguments{
\item{object}{A life table object}
}
\value{
a function coresponding to the CDF that defined the life table. If you'r life table is multidimentional, so will be the CDF.
}
\description{
This method gets you the cdf coresponding to the life table passed as the \code{object} parameter. Be carefull, because life tables could be defined in a discrete way, and in this case this method wont work, but you could rely on the underlying data.frame that you can get via \code{\link{getTable}}.
}
\examples{
getCdf(expolaw)(10)
getCdf(expolaw2)(30)
\dontrun{
getCdf(th002d)(5)
}
}
\seealso{
\code{\link{Lx}}, \code{\link{getAgeMax}}, \code{\linkS4class{TableMortContinue}}, \code{\link{getTable}}
}
