% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{getName}
\alias{getName}
\title{\code{getName} -- Get the name of a life table}
\usage{
getName(object)
}
\arguments{
\item{object}{A life table object}
}
\value{
the value of the slot "name" of the lifetable.
}
\description{
\code{getName} -- Get the name of a life table
}
\examples{
getName(th002d)
getName(tf002v)
getName(expolaw)
}
\seealso{
\code{\link{TableMort}}
}
