% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{Lx}
\alias{Lx}
\title{\code{Lx} -- Get the number of living people at some age.}
\usage{
Lx(table, x, UDD = TRUE)
}
\arguments{
\item{table}{A life table object}

\item{x}{A numerical age (or a vector or a matrix) representing the age of the person.}

\item{UDD}{A boolean -- Should UDD been used ?}
}
\value{
a numeric coreponding to maximum ages that insuree could be.
}
\description{
This method gets you the maximum age from a given life table. The life table have to be discrete. 
If UDD is given, the method can use inter-ages \code{x} paramters. This method is fully vectorized, as you can see in following exemples.
}
\examples{
Lx(expolaw, 12.7)
Lx(expolaw2, 12.7)
Lx(th002d, 12.7)
Lx(th002v, 12, UDD = FALSE)
Lx(tf002v, 12.7, UDD = TRUE)
Lx(tf002d, 12.7, UDD = TRUE)

#The Lx method is properly vectorised as follows : 
Lx(tableGPDId, c(1, 2, 3, 4))
Lx(tableGPDId, c(1, 2, 3, 4))
Lx(tableGPDId, matrix(1:12, nrow = 4))
Lx(tableGPDId, cbind(c(1, 2, 3, 4), c(1, 3, 4, 5), c(1, 5, 7, 9)))


}
\seealso{
\code{\link{Lx}}, \code{\link{getAgeMax}}, \code{\linkS4class{TableMortDiscrete}}, \code{\link{getTable}}
}
