% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{qxt}
\alias{qxt}
\alias{pxt}
\title{\code{pxt}-- Probabilities of survival and death.}
\usage{
qxt(table, x, t = 1)

pxt(table, x, t = 1)
}
\arguments{
\item{table}{A \code{\link{TableMort}} object.}

\item{x}{A numerical age (or a vector or a matrix) representing the age of the person}

\item{t}{Defaulted to 1, represent the time that you want to person to survive (for pxt) or to die in (for qxt)}
}
\value{
A numeric between 0 and 1, the probability of the event that a person age \code{x}, given the life table \code{table} survives for \code{t} years at least (\code{pxt}) or dies in the next \code{t} years (\code{qxt}).
}
\description{
The two methods \code{qxt} and \code{pxt} allow us to get probabilities of survival and death from life tables.
Note that \code{qxt} is just a wrapper for 1 - pxt.
}
\examples{

pxt(tableGPDIc, x = c(1, 2, 3, 4), t = c(1, 2, 3))
pxt(tableGPDIc, x = c(1, 2, 3, 4), t = c(1, 2, 3))
pxt(tableGPDIc, x = cbind(c(3, 4, 5, 6), c(1, 2, 3, 4)), t = c(1, 2, 3))
pxt(tableGPDId, x = c(1, 2, 3, 4), t = c(1, 2, 3))
pxt(tableGPDId, x = cbind(c(3, 4, 5, 6), c(1, 2, 3, 4)), t = c(1, 2, 3))

# Test the sum -- Should be one : 
pxt(th002d,1,5) + qxt(th002d,1,5) = 1
}
\seealso{
\code{\link{Lx}}, \code{\link{getAgeMax}}, \code{\linkS4class{TableMort}}, \code{\link{getTable}}
}
