% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verbosity.R
\name{setVerbosity}
\alias{setVerbosity}
\title{Set verbosity lvl}
\usage{
setVerbosity(x)
}
\arguments{
\item{x}{An integer, 0, 1 or 2.}
}
\value{
Nothing. 0 to set normal verbosity, 1 to talk a little more and 2 for
  debug.
}
\description{
\code{setVerbosity} allow you to set the verbosity lvl of life insurance
classes and life table classes and everything that comes out of actulife
package.
}
\examples{
setVerbosity(0) # default value
setVerbosity(1) # much more talking
setVerbosity(2) # debug mode, talks a loot.
}
