% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableMort.R
\docType{class}
\name{TableMort}
\alias{TableMort}
\title{\code{TableMort} - Mortality tables virtual motherclass.}
\description{
This class is a virtual class that's used to bind together all mortality table classes ( Continuous, discretes, groups...) that come with \code{actulife} package.
The Life tables classes that are availiables are \code{\link{TableMortDiscrete}}, \code{\link{TableMortContinue}}, \code{\link{TableMortGroupePDI}}, \code{\link{TableMortGroupePDIDiscrete}} and \code{\link{TableMortGroupePDIContinue}}. Depending on the kind of life table that you'r working with, you might use the Discrete's ones for discrete (whole year) life table an calculations, or the COntinuous one's for continuous life tables or calculations. You might also use the GroupePDI ( wich stands for "First death independant group") if you'r calculations deals with more than one life. Last but not least, the class \code{\link{Assure}} uses all of them in a nice and userfriendly approach. If you dont know how to start, you should go there.   
The following methods are availiable for all mortality tables as member of this motherclass :
}
\details{
\code{\link{getName}} -- Simple method to get the name of the mortality table.
\code{\link{pxt}} -- Probability of survival
\code{\link{qxt}} -- Probability of death
\code{\link{Ext}} -- Pure endowment
\code{\link{ex}} -- Expectations of life (curtate or not)
\code{\link{ax}} -- anuities (due, imediate, whole life or not, EoY, EoM, continuous, etc.)
\code{\link{Ax}} -- Insurences (same parametrisations as anuities)
\code{\link{summary}} -- Gives you a summary of the life table
}
